+++ mmgen/models/losses/__init__.py
@@ -5,10 +5,11 @@ from .disc_auxiliary_loss import (DiscShiftLoss, GradientPenaltyLoss,
                                   r1_gradient_penalty_loss)
 from .gan_loss import GANLoss
 from .gen_auxiliary_loss import GeneratorPathRegularizer, gen_path_regularizer
-from .pixelwise_loss import L1Loss, MSELoss
+from .pixelwise_loss import L1Loss, MSELoss, gaussian_kld
 
 __all__ = [
     'GANLoss', 'DiscShiftLoss', 'disc_shift_loss', 'gradient_penalty_loss',
     'GradientPenaltyLoss', 'R1GradientPenalty', 'r1_gradient_penalty_loss',
-    'GeneratorPathRegularizer', 'gen_path_regularizer', 'MSELoss', 'L1Loss'
+    'GeneratorPathRegularizer', 'gen_path_regularizer', 'MSELoss', 'L1Loss',
+    'gaussian_kld'
 ]
