+++ tests/test_cores/test_metrics.py
@@ -66,69 +66,6 @@ def test_ms_ssim():
     assert ssim_result < 1
 
 
-def test_kld_gaussian():
-    # we only test at bz = 1 to test the numerical accuracy
-    # due to the time and memory cost
-    tar_shape = [2, 3, 4, 4]
-    mean1, mean2 = torch.rand(*tar_shape, 1), torch.rand(*tar_shape, 1)
-    # var1, var2 = torch.rand(2, 3, 4, 4, 1), torch.rand(2, 3, 4, 4, 1)
-    var1 = torch.randint(1, 3, (*tar_shape, 1))
-    var2 = torch.randint(1, 3, (*tar_shape, 1))
-
-    def pdf(x, mean, var):
-        return (1 / np.sqrt(2 * np.pi * var) * torch.exp(-(x - mean)**2 /
-                                                         (2 * var)))
-
-    delta = 0.0001
-    indice = torch.arange(-5, 5, delta).repeat(*mean1.shape)
-    p = pdf(indice, mean1, var1)  # pdf of target distribution
-    q = pdf(indice, mean2, var2)  # pdf of predicted distribution
-
-    kld_manually = (p * torch.log(p / q) * delta).sum(dim=(1, 2, 3, 4)).mean()
-
-    data = dict(
-        mean_pred=mean2,
-        mean_target=mean1,
-        logvar_pred=torch.log(var2),
-        logvar_target=torch.log(var1))
-
-    metric = GaussianKLD(2)
-    metric.prepare()
-    metric.feed(data, 'reals')
-    kld = metric.summary()
-    # this is a quite loose limitation for we cannot choose delta which is
-    # small enough for precise kld calculation
-    np.testing.assert_almost_equal(kld, kld_manually, decimal=1)
-    # assert (kld - kld_manually < 1e-1).all()
-
-    metric_base_2 = GaussianKLD(2, log_base='2')
-    metric_base_2.prepare()
-    metric_base_2.feed(data, 'reals')
-    kld_base_2 = metric_base_2.summary()
-    np.testing.assert_almost_equal(kld_base_2, kld / np.log(2), decimal=4)
-    # assert kld_base_2 == kld / np.log(2)
-
-    # test wrong log_base
-    with pytest.raises(AssertionError):
-        GaussianKLD(2, log_base='10')
-
-    # test other reduction --> mean
-    metric = GaussianKLD(2, reduction='mean')
-    metric.prepare()
-    metric.feed(data, 'reals')
-    kld = metric.summary()
-
-    # test other reduction --> sum
-    metric = GaussianKLD(2, reduction='sum')
-    metric.prepare()
-    metric.feed(data, 'reals')
-    kld = metric.summary()
-
-    # test other reduction --> error
-    with pytest.raises(AssertionError):
-        metric = GaussianKLD(2, reduction='none')
-
-
 class TestExtractInceptionFeat:
 
     @classmethod
@@ -162,9 +99,8 @@ class TestExtractInceptionFeat:
 
     @pytest.mark.skipif(not torch.cuda.is_available(), reason='requires cuda')
     def test_extr_inception_feat_cuda(self):
-        # inception = torch.nn.DataParallel(self.inception)
-        feat = extract_inception_features(self.data_loader,
-                                          self.inception.cuda(), 5)
+        inception = torch.nn.DataParallel(self.inception)
+        feat = extract_inception_features(self.data_loader, inception, 5)
         assert feat.shape[0] == 5
 
     @torch.no_grad()
@@ -178,8 +114,8 @@ class TestExtractInceptionFeat:
         # Tero implementation
         net = torch.jit.load(
             './work_dirs/cache/inception-2015-12-05.pt').eval().cuda()
-        # net = torch.nn.DataParallel(net)
-        feature_tero = net(img.cuda(), return_features=True)
+        net = torch.nn.DataParallel(net)
+        feature_tero = net(img, return_features=True)
 
         print(feature_ours.shape)
 
@@ -328,3 +264,66 @@ class TestPPL:
             ppl.feed(b, 'fakes')
         score = ppl.summary()
         assert score >= 0
+
+
+def test_kld_gaussian():
+    # we only test at bz = 1 to test the numerical accuracy
+    # due to the time and memory cost
+    tar_shape = [2, 3, 4, 4]
+    mean1, mean2 = torch.rand(*tar_shape, 1), torch.rand(*tar_shape, 1)
+    # var1, var2 = torch.rand(2, 3, 4, 4, 1), torch.rand(2, 3, 4, 4, 1)
+    var1 = torch.randint(1, 3, (*tar_shape, 1))
+    var2 = torch.randint(1, 3, (*tar_shape, 1))
+
+    def pdf(x, mean, var):
+        return (1 / np.sqrt(2 * np.pi * var) * torch.exp(-(x - mean)**2 /
+                                                         (2 * var)))
+
+    delta = 0.0001
+    indice = torch.arange(-5, 5, delta).repeat(*mean1.shape)
+    p = pdf(indice, mean1, var1)  # pdf of target distribution
+    q = pdf(indice, mean2, var2)  # pdf of predicted distribution
+
+    kld_manually = (p * torch.log(p / q) * delta).sum(dim=(1, 2, 3, 4)).mean()
+
+    data = dict(
+        mean_pred=mean2,
+        mean_target=mean1,
+        logvar_pred=torch.log(var2),
+        logvar_target=torch.log(var1))
+
+    metric = GaussianKLD(2)
+    metric.prepare()
+    metric.feed(data, 'reals')
+    kld = metric.summary()
+    # this is a quite loose limitation for we cannot choose delta which is
+    # small enough for precise kld calculation
+    np.testing.assert_almost_equal(kld, kld_manually, decimal=1)
+    # assert (kld - kld_manually < 1e-1).all()
+
+    metric_base_2 = GaussianKLD(2, base='2')
+    metric_base_2.prepare()
+    metric_base_2.feed(data, 'reals')
+    kld_base_2 = metric_base_2.summary()
+    np.testing.assert_almost_equal(kld_base_2, kld / np.log(2), decimal=4)
+    # assert kld_base_2 == kld / np.log(2)
+
+    # test wrong log_base
+    with pytest.raises(AssertionError):
+        GaussianKLD(2, base='10')
+
+    # test other reduction --> mean
+    metric = GaussianKLD(2, reduction='mean')
+    metric.prepare()
+    metric.feed(data, 'reals')
+    kld = metric.summary()
+
+    # test other reduction --> sum
+    metric = GaussianKLD(2, reduction='sum')
+    metric.prepare()
+    metric.feed(data, 'reals')
+    kld = metric.summary()
+
+    # test other reduction --> error
+    with pytest.raises(AssertionError):
+        metric = GaussianKLD(2, reduction='none')
